Extract shared infrastructure into server/ modules
Move fetch helpers (secureFetch, fetchJSON, isPrivateIP), cache eviction,
Cache-Control middleware, and security/bootstrap setup out of server.js
into dedicated modules under server/services/ and server/middleware/.
No endpoint or response behavior changes — server.js delegates to imports.
